Move `RelationshipIds` to panel v2 only
