+++ libs/designer/src/lib/core/parsers/pasteScopeInWorkflow.ts
@@ -1,4 +1,4 @@
-import type { RelationshipIds } from '../state/panel/panelInterfaces';
+import type { RelationshipIds } from '../state/panelV2/panelTypes';
 import type { NodesMetadata, Operations, WorkflowState } from '../state/workflow/workflowInterfaces';
 import type { WorkflowNode } from './models/workflowNode';
 import { addNewEdge, reassignEdgeSources, reassignEdgeTargets, removeEdge, applyIsRootNode } from './restructuringHelpers';
